store_to_map: cast next_token_ids to int64 (slice->advanced index lost implicit cast)

+++ python/sglang/srt/managers/overlap_utils.py
@@ -122,7 +122,11 @@ class FutureMap:
         self, future_indices: FutureIndices, batch_result: GenerationBatchResult
     ):
         if self.spec_algo.is_none():
-            self.token_ids_buf[future_indices.indices] = batch_result.next_token_ids
+            # next_token_ids is int32; buf is int64. Slice assignment used to
+            # cast implicitly, but advanced indexing requires an explicit match.
+            self.token_ids_buf[future_indices.indices] = batch_result.next_token_ids.to(
+                torch.int64
+            )
         else:
             draft_input: EagleDraftInput = batch_result.next_draft_input
             self.store_to_map_for_new_batch(future_indices, draft_input)
